Add support for ARCH=arm
Add support for building with ARCH=arm.

The uClibc config is a slightly hand-tweaked version of the defconfig;
in particular it disables shared library support and uses little endian
rather than big endian, as well as fixing various non-arch-specific
options to match the x86 config.

The kernel config was produced by
 mkdir tmpk && cd tmpk
 make -C $(pwd)/../linux O=$(pwd) ARCH=arm vexpress_defconfig

The resulting kernel and initrd can be run with
 arm-softmmu/qemu-system-arm -kernel build-arm/linux/arch/arm/boot/zImage \
 -initrd build-arm/initramfs.img -M vexpress-a9 -serial stdio -m 1024 \
 -append 'mem=1024M console=ttyAMA0,38400n8'

Note that something in the build/install process creates an empty
directory "install-arm/-linux-uclibc/lib/" but this seems to be harmless.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>